+++ lib/framework/widgets/user_widgets.dart
@@ -116,57 +116,3 @@ class UserWidgets {
     );
   }
 }
-
-class UserProfileCard extends StatelessWidget {
-  final String userName;
-  final String userImage;
-  final String userProfileId;
-
-  const UserProfileCard({
-    super.key,
-    required this.userName,
-    required this.userImage,
-    required this.userProfileId,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: WidgetsSizes.profileTabHeight,
-      padding: EdgeInsets.only(left: Sizes.size20),
-      decoration: BoxDecoration(
-        color: CustomColors.mainBackgroundColor,
-      ),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          CircleAvatar(
-            backgroundImage: AssetImage(userImage),
-            radius: Sizes.size22.sp,
-            backgroundColor: Colors.transparent,
-          ),
-          SizedBox(width: Sizes.size15),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                TextWidget(
-                  text: userName,
-                  color: Colors.white,
-                  size: Sizes.size18.sp,
-                  type: FontWeight.bold,
-                ),
-                TextWidget(
-                  text: '@$userProfileId',
-                  color: CustomColors.lightBackgroundColor,
-                  size: Sizes.size16.sp,
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
